Re-Organize Main Controller code
