+++ pkg/controller/cluster/log-search.go
@@ -15,8 +15,16 @@
 package cluster
 
 import (
+	"context"
 	"errors"
 	"fmt"
+	"net/http"
+	"time"
+
+	"github.com/minio/madmin-go"
+	"github.com/minio/operator/pkg/resources/secrets"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	miniov2 "github.com/minio/operator/pkg/apis/minio.min.io/v2"
 	"github.com/minio/operator/pkg/resources/deployments"
@@ -89,4 +97,209 @@ func logSearchAPIDeploymentMatchesSpec(tenant *miniov2.Tenant, actualDeployment 
 		return false, nil
 	}
 	return true, nil
+}
+
+func (c *Controller) deleteLogHeadlessService(ctx context.Context, tenant *miniov2.Tenant) error {
+	_, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.LogHLServiceName())
+	if k8serrors.IsNotFound(err) {
+		return nil
+	}
+	klog.V(2).Infof("Deleting Log Headless Service for %s", tenant.Namespace)
+	err = c.kubeClientSet.CoreV1().Services(tenant.Namespace).Delete(ctx, tenant.LogHLServiceName(), metav1.DeleteOptions{})
+	c.RegisterEvent(ctx, tenant, corev1.EventTypeWarning, "Deleted", "Log search headless service deleted")
+	return err
+}
+
+func (c *Controller) deleteLogStatefulSet(ctx context.Context, tenant *miniov2.Tenant) error {
+	_, err := c.statefulSetLister.StatefulSets(tenant.Namespace).Get(tenant.LogStatefulsetName())
+	if k8serrors.IsNotFound(err) {
+		return nil
+	}
+	klog.V(2).Infof("Deleting Log StatefulSet for %s", tenant.Namespace)
+	err = c.kubeClientSet.AppsV1().StatefulSets(tenant.Namespace).Delete(ctx, tenant.LogStatefulsetName(), metav1.DeleteOptions{})
+	c.RegisterEvent(ctx, tenant, corev1.EventTypeWarning, "Deleted", "Log search statefulset deleted")
+	return err
+}
+
+func (c *Controller) deleteLogSearchAPIDeployment(ctx context.Context, tenant *miniov2.Tenant) error {
+	_, err := c.deploymentLister.Deployments(tenant.Namespace).Get(tenant.LogSearchAPIDeploymentName())
+	if k8serrors.IsNotFound(err) {
+		return nil
+	}
+	klog.V(2).Infof("Deleting Log Search API deployment for %s", tenant.Name)
+	err = c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Delete(ctx, tenant.LogSearchAPIDeploymentName(), metav1.DeleteOptions{})
+	c.RegisterEvent(ctx, tenant, corev1.EventTypeWarning, "Deleted", "Log search deployment deleted")
+	return err
+}
+
+func (c *Controller) deleteLogSearchAPIService(ctx context.Context, tenant *miniov2.Tenant) error {
+	_, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.LogSearchAPIServiceName())
+	if k8serrors.IsNotFound(err) {
+		return nil
+	}
+	klog.V(2).Infof("Delete Log Search API Service for %s", tenant.Namespace)
+	err = c.kubeClientSet.CoreV1().Services(tenant.Namespace).Delete(ctx, tenant.LogSearchAPIServiceName(), metav1.DeleteOptions{})
+	c.RegisterEvent(ctx, tenant, corev1.EventTypeWarning, "Deleted", "Log search service deleted")
+	return err
+}
+
+func (c *Controller) checkAndCreateLogHeadless(ctx context.Context, tenant *miniov2.Tenant) (*corev1.Service, error) {
+	svc, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.LogHLServiceName())
+	if err == nil || !k8serrors.IsNotFound(err) {
+		return svc, err
+	}
+
+	klog.V(2).Infof("Creating a new Log Headless Service for %s", tenant.Namespace)
+	svc = services.NewHeadlessForLog(tenant)
+	_, err = c.kubeClientSet.CoreV1().Services(svc.Namespace).Create(ctx, svc, metav1.CreateOptions{})
+	c.RegisterEvent(ctx, tenant, corev1.EventTypeNormal, "SvcCreated", "Log headless service created")
+	return svc, err
+}
+
+func (c *Controller) checkAndCreateLogStatefulSet(ctx context.Context, tenant *miniov2.Tenant, svcName string) error {
+	logPgSS, err := c.statefulSetLister.StatefulSets(tenant.Namespace).Get(tenant.LogStatefulsetName())
+	if err != nil {
+		if !k8serrors.IsNotFound(err) {
+			return err
+		}
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusProvisioningLogPGStatefulSet, 0); err != nil {
+			return err
+		}
+
+		klog.V(2).Infof("Creating a new Log StatefulSet for %s", tenant.Namespace)
+		searchSS := statefulsets.NewForLogDb(tenant, svcName)
+		_, err = c.kubeClientSet.AppsV1().StatefulSets(tenant.Namespace).Create(ctx, searchSS, metav1.CreateOptions{})
+		c.RegisterEvent(ctx, tenant, corev1.EventTypeNormal, "Created", "Log statefulset created")
+		return err
+
+	}
+
+	// check if expected and actual values of Log DB spec match
+	dbSpecMatches, err := logDBStatefulsetMatchesSpec(tenant, logPgSS)
+	if err != nil {
+		return err
+	}
+	if !dbSpecMatches {
+		// Note: using current spec replica count works as long as we don't expose replicas via tenant spec.
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusUpdatingLogPGStatefulSet, *logPgSS.Spec.Replicas); err != nil {
+			return err
+		}
+		logPgSS = statefulsets.NewForLogDb(tenant, svcName)
+		if _, err = c.kubeClientSet.AppsV1().StatefulSets(tenant.Namespace).Update(ctx, logPgSS, metav1.UpdateOptions{}); err != nil {
+			return err
+		}
+		c.RegisterEvent(ctx, tenant, corev1.EventTypeNormal, "Updated", "Log statefulset updated")
+	}
+
+	return nil
+}
+
+func (c *Controller) checkAndCreateLogSearchAPIService(ctx context.Context, tenant *miniov2.Tenant) error {
+	_, err := c.serviceLister.Services(tenant.Namespace).Get(tenant.LogSearchAPIServiceName())
+	if err == nil || !k8serrors.IsNotFound(err) {
+		return err
+	}
+
+	klog.V(2).Infof("Creating a new Log Search API Service for %s", tenant.Namespace)
+	svc := services.NewClusterIPForLogSearchAPI(tenant)
+	_, err = c.kubeClientSet.CoreV1().Services(svc.Namespace).Create(ctx, svc, metav1.CreateOptions{})
+	c.RegisterEvent(ctx, tenant, corev1.EventTypeNormal, "SvcCreated", "Log search service created")
+	return err
+}
+
+func (c *Controller) checkAndCreateLogSearchAPIDeployment(ctx context.Context, tenant *miniov2.Tenant) error {
+	logSearchDeployment, err := c.deploymentLister.Deployments(tenant.Namespace).Get(tenant.LogSearchAPIDeploymentName())
+	if err != nil {
+		if !k8serrors.IsNotFound(err) {
+			return err
+		}
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusProvisioningLogSearchAPIDeployment, 0); err != nil {
+			return err
+		}
+
+		klog.V(2).Infof("Creating a new Log Search API deployment for %s", tenant.Name)
+		_, err = c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Create(ctx, deployments.NewForLogSearchAPI(tenant), metav1.CreateOptions{})
+		c.RegisterEvent(ctx, tenant, corev1.EventTypeNormal, "Created", "Log search service deployment")
+		return err
+	}
+
+	// check if expected and actual values of Log search API deployment match
+	apiDeploymentMatches, err := logSearchAPIDeploymentMatchesSpec(tenant, logSearchDeployment)
+	if err != nil {
+		return err
+	}
+	if !apiDeploymentMatches {
+		// Note: using current spec replica count works as long as we don't expose replicas via tenant spec.
+		if tenant, err = c.updateTenantStatus(ctx, tenant, StatusUpdatingLogSearchAPIServer, 0); err != nil {
+			return err
+		}
+		logSearchDeployment = deployments.NewForLogSearchAPI(tenant)
+		if _, err := c.kubeClientSet.AppsV1().Deployments(tenant.Namespace).Update(ctx, logSearchDeployment, metav1.UpdateOptions{}); err != nil {
+			return err
+		}
+		c.RegisterEvent(ctx, tenant, corev1.EventTypeNormal, "Updated", "Log search deployment updated")
+	}
+	return nil
+}
+
+func (c *Controller) checkAndCreateLogSecret(ctx context.Context, tenant *miniov2.Tenant) (*corev1.Secret, error) {
+	secret, err := c.kubeClientSet.CoreV1().Secrets(tenant.Namespace).Get(ctx, tenant.LogSecretName(), metav1.GetOptions{})
+	if err == nil || !k8serrors.IsNotFound(err) {
+		return secret, err
+	}
+
+	klog.V(2).Infof("Creating a new Log secret for %s", tenant.Name)
+	secret, err = c.kubeClientSet.CoreV1().Secrets(tenant.Namespace).Create(ctx, secrets.LogSecret(tenant), metav1.CreateOptions{})
+	return secret, err
+}
+
+func (c *Controller) checkAndConfigureLogSearchAPI(ctx context.Context, tenant *miniov2.Tenant, secret *corev1.Secret, adminClnt *madmin.AdminClient) error {
+	// Check if audit webhook is configured for tenant's MinIO
+	auditCfg := newAuditWebhookConfig(tenant, secret)
+	_, err := adminClnt.GetConfigKV(ctx, auditCfg.target)
+	if err != nil {
+		// check if log search is ready
+		if err = c.checkLogSearchAPIReady(tenant); err != nil {
+			klog.V(2).Info(err)
+			if _, err = c.updateTenantStatus(ctx, tenant, StatusWaitingForLogSearchReadyState, 0); err != nil {
+				return err
+			}
+			return ErrLogSearchNotReady
+		}
+		restart, err := adminClnt.SetConfigKV(ctx, auditCfg.args)
+		if err != nil {
+			return err
+		}
+		if restart {
+			// Restart MinIO for config update to take effect
+			if err = adminClnt.ServiceRestart(ctx); err != nil {
+				klog.V(2).Info("error restarting minio")
+				klog.V(2).Info(err)
+			}
+			klog.V(2).Info("done restarting minio")
+		}
+		return nil
+	}
+	return err
+}
+
+func (c *Controller) checkLogSearchAPIReady(tenant *miniov2.Tenant) error {
+	endpoint := fmt.Sprintf("http://%s.%s.svc.%s:8080", tenant.LogSearchAPIServiceName(), tenant.Namespace, miniov2.GetClusterDomain())
+	client := http.Client{Timeout: 100 * time.Millisecond}
+	resp, err := client.Get(endpoint)
+	if err != nil {
+		fmt.Println(err)
+		return err
+	}
+	defer func() {
+		if err := resp.Body.Close(); err != nil {
+			klog.V(2).Info(err)
+		}
+	}()
+
+	if resp.StatusCode == 404 {
+		return nil
+	}
+
+	return errors.New("Log Search API Not Ready")
 }
